+++ pychron/lasers/pattern/pattern_maker_view.py
@@ -5,7 +5,7 @@
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
 #
-#   http://www.apache.org/licenses/LICENSE-2.0
+# http://www.apache.org/licenses/LICENSE-2.0
 #
 # Unless required by applicable law or agreed to in writing, software
 # distributed under the License is distributed on an "AS IS" BASIS,
@@ -40,9 +40,9 @@ class PatternMakerView(Saveable, Patternable):
                          'LineSpiral',
                          'SquareSpiral',
                          'Random',
-                         'CircularContour', 'Trough'),
+                         'CircularContour', 'Trough', 'Rubberband'),
                     depends_on='_kind')
-    _kind = Str('Trough')
+    _kind = Str('Rubberband')
 
     def load_pattern(self, path=None):
         if path is None:
@@ -93,6 +93,7 @@ class PatternMakerView(Saveable, Patternable):
                  show_label=False),
             handler=self.handler_klass,
             buttons=SaveableButtons,
+            height=425,
             title='Pattern Editor',
             resizable=True)
         return v